Refactor code - Move RamowkaInfo and RamowkaList to separate files - Move all components of Ramowka into separate directory - Add more clear doc comments - Change some variable names - Refactor _updateRamowka() method into multiple single-responsibility   methods
