+++ mobile/lib/resources/resources.dart
@@ -1,3 +1,4 @@
 export 'assets.gen.dart';
 export 'colors.dart';
 export 'theme.dart';
+export 'day.dart';
